test(RacingCar) : 자동차 경주 게임 개발 시작
- 테스트 케이스 수정 및 추가
